test: update test case
