+++ cli/tests/unit/opcall_test.ts
@@ -20,6 +20,13 @@ Deno.test(async function sendAsyncStackTrace() {
   }
 });
 
+declare global {
+  namespace Deno {
+    // deno-lint-ignore no-explicit-any, no-var
+    var core: any;
+  }
+}
+
 Deno.test(async function opsAsyncBadResource() {
   try {
     const nonExistingRid = 9999;
